Renamed DiscoverBrew to DiscoverBean for better naming. Implemented bean name to brew trak cards. Changed url to include bean/brew after discover
