+++ src/components/DiscoverBean/DiscoverCard.js
