+++ src/components/DiscoverBean/sampleDetails.js
