+++ src/queries/index.js
@@ -30,6 +30,7 @@ query get_recipes {
     date_added 
     bean {
       img
+      name
     }
   }
 }
